F3

+++ Viewer/include/Renderer.h
@@ -12,9 +12,9 @@ public:
 	void SwapBuffers();
 	void ClearColorBuffer(const glm::vec3& color);
 	int GetViewportWidth() const;
-	void SetViewportWidth(int New_width);
+	void SetViewportWidth(float New_width);
 	int GetViewportHeight() const;
-	void SetViewportHeight(int New_height);
+	void SetViewportHeight(float New_height);
 	
 private:
 	void PutPixel(const int i, const int j, const glm::vec3& color);
